requirements.txt
discord 
discord-webhook 
websocket-client
PyYAML
